include train in MLCommons class as  a func
